Improvements

Obs.v_imp — Function

 v_imp(vv,vt...;cv, L=1, theta1,theta2, bnd::Boundary=open)
 v_imp(vv::juobs.Corr,vt::juobs.Corr...;cv,L=1, theta1=Float64[],theta2=Float64[], bnd::Boundary=open)

Compute the improved G{VV} = \sum{k=1}^3G{ViV_i} correlator according to the equations

\[ G^I_{V_iV_i} = G_{V_i V_i} + 2ac_V \de_t G_{V_i T_i0} - 2c_V \sum_{j=1}^{3} i \sin(p^j a) G_{V_i T_ij}\]

Arguments

vv: correlator G_{VV}. It has to be already averaged over the three polarizations.

vt: correlators G{VT}. It should be either G{ViT0i} (already averaged over the three polarizations) if the correlator is at 0 momentum or G{ViT0i}, G{V1T12},G{V1T13}, G{V2T12},G{V2T23}, G{V3T13},G_{V3T23} if at non-zero momentum

Keyword Arguments

cv: mandatory keyword argument. It is the improvement coefficient

L::Int64=1: Lattice size. Needed to convert the theta angle into momenta.

theta1=Float64[]: theta angle of the first quark. If empty, is read from vv

theta2=Float64[]: theta angles of the second quark. If empty, is read from vv

bnd : Boundary condition of the lattice. Used to compute the derivatives

See also sym_der, Boundary

Obs.pv_imp — Function

 pv_imp(pv,pt...;cv,L::Int64=1,theta1, theta2, bnd::Boundary = open)
 pv_imp(pv::juobs.Corr, pt::juobs.Corr...; cv, L=1, theta1=Float64[],theta2 = Float64[], bnd::Boundary=open)

Compute the improved G{PV} = 1/3 \sum{k=1}^3G{PVi} correlator according to the equations

\[ G^I_{PV_i} = G_{P V_i} + ac_V \de_t G_{P T_i0} - c_V \sum_{j=1}^{3} \sin(p^j a) G_{P T_ij}\]

Arguments

pv: correlator G_{PV}. It has to be already averaged over the three polarizations.

pt: correlators G{PT}. It should be either G{PT0i} (already averaged over the three polarizations) if the correlator is at 0 momentum or G{PT0i}, G{PT12},G{PT13},G{PT23} if at non-zero momentum

Keyword Arguments

cv: mandatory keyword argument. It is the improvement coefficient

L::Int64=1: Lattice size. Needed to convert the theta angle into momenta.

theta1=Float64[]: theta angle of the first quark. If empty, is read from pv

theta2=Float64[]: theta angles of the second quark. If empty, is read from pv

bnd : Boundary condition of the lattice. Used to compute the derivatives

See also sym_der, Boundary

Obs.pv0_imp — Function

 pv0_imp(pv0, pt...; theta1,theta2, cv,L::Int64=1, bnd::Boundary=open)
 pv0_imp(pv0::juobs.Corr,pt::juobs.Corr ...;cv,L=1,theta1 = Float64[], theta2 = Float64[], bnd::Boundary=open)::juobs.Corr

Improve the correlator GPV0 and GPV0 with the tensor correlator according to the improvement equation

\[ V_0^I(t,\vec p) = V_0(t,\vec p) - c_V \sum_{k=1}^3 i sin(ap^k) T_{0k}(t,\vec p)\]

Arguments

pvo: correlator G_{PV0}

pt: correlators G{PT}. It contains the correlator G{PT01}, G{PT02}, G{PT03}

Keyword Arguments

cv: mandatory keyword argument. It is the improvement coefficient

L::Int64=1: Lattice size. Needed to convert the theta angle into momenta.

theta1=Float64[]: theta angle of the first quark. If empty, is read from pv

theta2=Float64[]: theta angles of the second quark. If empty, is read from pv

bnd : Boundary condition of the lattice. Used to compute the derivatives

See also sym_der, Boundary

Obs.pa0_imp — Function

pa0_imp(pa0,pp; ca, bnd::Boundary =open)
pa0_imp(pa0::juobs.Corr,pp::juobs.Corr; ca, bnd::Boundary =open)::juobs.Corr

Improve the correlator G{PA0} according to the equation

\[ G_{PA_0^{I}} = G_{PA_0} - a c_A \\de_t G_{PP}\]

Keyword Arguments

ca: mandatory keyword argument. It is the improved operator [See also ca_fit]

bnd : Boundary condition of the lattice. Used to compute the derivatives

See also sym_der, Boundary

Obs.a0a0_imp — Function

a0a0_imp(a0a0,pa0; ca,  bnd::Boundary=open)
a0a0_imp(a0a0::juobs.Corr,pa0::juobs.Corr; ca,  bnd::Boundary=open)::juobs.Corr

Improve the correlator G{A0 A_0} according to the equation

\[ G_{A_0^{I} A_0^{I}} = G_{A_0 A_0} - 2 a c_A \de_t G_{P A_0}\]

Keyword Arguments

ca: mandatory keyword argument. It is the improved operator [See also ca_fit]

bnd : Boundary condition of the lattice. Used to compute the derivatives

See also sym_der, Boundary