Observables

Obs.mpcac — Function

mpcac(a0p,pp [,ca], bnd::Boundary = open)

Computes the effective PCAC mass associated to the correlators a0p and a pp according to the boundary condition bnd. If ca is given, it first improve the a0p correlator according to a0p-ca der_pp and then computes the pcac mass. It uses the symmetryc derivatives where necessary.

Output specifics

If a0p and/or pp are juobs.Corr, then according to bnd the code uses:

bnd == open: the obs field where it strips the first and last component. This because in OBC the first and last component lay on the boundary and are 0

bnd == periodic: the obs field without modification

See also sym_der, Boundary, pa0_imp

Obs.meff — Function

 meff(v,bnd::Boundary=open)

Compute the effective mass associated to the correlator v according to the boundary condition bnd.

Output Specifics

The length of the retuned vector is length(v)-1 if OBC, while is length(v) if PBC

Obs.ps_dec — Function

 ps_dec(pa0,pp,mps,y0)

It computes the effective pseudoscalar decay constant associated to the correlator pa0 and pp. mps is the pseudoscalar mass and y0 is the source position.

If pa0 and pp are a juobs.Corr, y0 can be omitted and it will take the source position in pa0

Obs.dec — Function

 dec(c,m,y0)

It computes the effective decay constant associated to the correlator c. m is the mass associated to it and y0 the source position.

If c is a juobs.Corr, y0 can be omitted and it will take the source position in c

Obs.RI — Function

RI(HtoL, H, L, EL, EH; xsink, xsrc)

Compute the matrix element associated to the semileptonic decay H -> L using the ratio

\[ R_I(t) = \sqrt(2E_L) HtoL(t)/\sqrt(H(x_{sink}-t) L(t)) \times \exp(\frac{1}{2} (E_L t + E_H(x_{sink}-t)))\]

Obs.RII — Function

RII(HtoL, H, L, EL, EH; xsink, xsrc)

Compute the matrix element associated to the semileptonic decay H -> L using the ratio

\[ R_{II}^2(t) = 2E_L \frac{HtoL(t)HtoL(x_{sink}-t)}{H(x_{sink})L(x_{sink})}\]

Obs.RIII — Function

RIII(HtoL, H, L, EL, EH; xsink, xsrc)

Compute the matrix element associated to the semileptonic decay H -> L using the ratio from 1903.05870 eq 5.1

\[ R_{III}(t) = \sqrt(2E_L) HtoL(t)/\sqrt(H(x_{sink}) L(x_{sink})) \times \exp(\frac{x_{sink}-2t}{2} (E_H - E_L))\]