Utilities

Obs.sym_der — Function

 sym_der(v,bnd::Boundary)

Compute the symmetryc derivative of v according to the boundary condition in bnd See also Boundary

Obs.check_corr — Function

 check_corr(c::juobs.Corr...; flag::Check_flag)

check that all correlator are compatible. Use flag to ignore specific fields

See also Check_flag

Obs.average_corr — Function

 average_corr(x::juobs.Corr...; flag::Check_flag = no_gamma)

It average the correlator x.... It first check for compatibility among the correlator according to flag. It is meant to average correlator with vector and tensor currents and flag = no_gamma.

Output Specifics

The function returns a juobs.Corr with gamma structure made by appending the gamma structure of the original correlator, so that if we average three correlator with gamma structure G1, G2 and G3, the returned correlator has gamma structure G1,G2,G3.

#TODO: if flags is something different that no_gamma, the correlators will be averaged correctly but other informations will be lost (:kappa, :mu, ecc...)

See also check_corr, Check_flag

Obs.sym_source — Function

sym_source(corr, y0, parity, bnd::Boundary = open)
sym_source(corr::juobs.Corr [,y0], parity, bnd::Boundary = open)

symmetrize the correlator with respect to the source position y0

Arguments

y0 source position in Julia indexing.

parity parity of the correlator. should be \pm 1

bnd the boundary condition of the correlator

Return types

(1) returns a vector of the same type as corr

(2) returns a juobs.Corr

See also Boundary